+++ docs/changelog.md
@@ -75,6 +75,9 @@ As features stabilize some brief notes about them will accumulate here.
   easier to spot the remaining candidates. Thanks to @mr-felixoid and @bew! #7752
 
 #### New
+* [OpenReviewPane](config/lua/keyassignment/OpenReviewPane.md) opens a git diff
+  review pane: navigate hunks/files, annotate lines, and send selected diff
+  lines to another pane.
 * [wezterm.serde](config/lua/wezterm.serde/index.md) module for serialization
   and deserialization of JSON, TOML and YAML. Thanks to @expnn! #4969
 * `wezterm ssh` now supports agent forwarding. Thanks to @Riatre! #5345